\name{SAGA-package}
\alias{SAGA-package}
\alias{SAGA}
\docType{package}
\title{
Software for the Analysis of Genetic Architecture

SAGA}
\description{
Provides multimodel parameter estimates of genetic architecture.  Uses 
weighted least squares to solve all possible models.  Chooses the model or 
models that are most appropriate based on AICc weights and produces model 
weighted parameter estimates and unconditional standard errors.  A function to visualize 
the distribution of model fits across model space allows users to graphically 
see how well defined the "true" model is with their dataset.
}
\details{
\tabular{ll}{
Package: \tab SAGA\cr
Version: \tab 2.0.0\cr
}
}
\author{
Heath Blackmon and Jeffery P. Demuth\cr
coleoguy@gmail.com}
\references{
Lynch, M. and B. Walsh. 1998. Genetics and analysis of quantitative traits. Sinauer Associates, Inc., Sunderland, Massachusetts.\cr\cr
Burnham, K. P. and D. R. Anderson. 2002. Model selection and multimodel inference: a practical information-theoretic approach. Springer Science & Business Media.}
